removed unused files, restructuring
