+++ engine/output_generator_flaskserver_socket_io.py
@@ -1,6 +1,6 @@
 from flask import stream_with_context, request, Response, Flask
 from realtime_VideoStreamer import VideoStreamer
-from realtime_RecognitionEngine_textOutput_v2 import RecognitionEngine
+from realtime_RecognitionEngine_textOutput import RecognitionEngine
 from keras.models import load_model
 from flask_socketio import SocketIO, emit
