--- engine/sandbbox.py
@@ -1,3 +1,0 @@
-links = "https://www.twitch.tv/a541021,https://www.twitch.tv/a541021,https://www.twitch.tv/a541021,https://www.twitch.tv/a541021"
-
-print(links.split(","))
